# Optimisation: Avoid set copies when compiling files