--- a/src/fluent_compiler/codegen.py
+++ b/src/fluent_compiler/codegen.py
@@ -116,12 +116,6 @@ class Scope:
         self._properties = {}
         self._assignments = {}
 
-    def names_in_use(self):
-        names = self.names
-        if self.parent_scope is not None:
-            names = names | self.parent_scope.names_in_use()
-        return names
-
     def is_name_in_use(self, name: str) -> bool:
         if name in self.names:
             return True
@@ -131,12 +125,6 @@ class Scope:
 
         return self.parent_scope.is_name_in_use(name)
 
-    def function_arg_reserved_names(self):
-        names = self._function_arg_reserved_names
-        if self.parent_scope is not None:
-            names = names | self.parent_scope.function_arg_reserved_names()
-        return names
-
     def is_name_reserved_function_arg(self, name: str) -> bool:
         if name in self._function_arg_reserved_names:
             return True
@@ -145,9 +133,6 @@ class Scope:
             return False
 
         return self.parent_scope.is_name_reserved_function_arg(name)
-
-    def all_reserved_names(self):
-        return self.names_in_use() | self.function_arg_reserved_names()
 
     def is_name_reserved(self, name: str) -> bool:
         return self.is_name_in_use(name) or self.is_name_reserved_function_arg(name)