@spookylukey 2024-05-30
This looks great, thanks!

On my Linux box I'm seeing even bigger speedups, from ~28s to ~6s on the same benchmark.